+++ src/lib/Components/Bidding/Components/Input.tsx
@@ -19,6 +19,12 @@ export class Input extends Component<InputProps, InputState> {
     }
   }
 
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.error) {
+      this.setState({ borderColor: "red100" })
+    }
+  }
+
   onBlur() {
     this.setState({
       borderColor: this.props.error ? "red100" : "black10",