+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Adds a checkbox to the confirm your bid screen - yuki24
 * Adds the ability to style the text in the inverted button with a spinner - yuki24
 * Now the `<ConfirmBid>` component shows a spinner while loading - yuki24
+* Now the `<Input>` component updates the border color when the parent component updates the `error` prop - yuki24
 
 ### 1.5.1